+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+next release
+------
+
+* Add opensearch support providing ClientBuilder to all elasticsearch checks
+
 v1.3.6
 ------
 
@@ -168,4 +173,4 @@ Add next checks:
 * `\FiveLab\Component\Diagnostic\Check\Eloquent\DatabaseConnectionCheck` - check db connection via Eloquent.
 * `\FiveLab\Component\Diagnostic\Check\Eloquent\DatabaseMysqlVersionCheck` - check MySQL version via Eloquent.
 * `\FiveLab\Component\Diagnostic\Check\Elasticsearch\ElasticsearchClusterStateCheck` - check ES cluster health check.
-* `\FiveLab\Component\Diagnostic\Check\Elasticsearch\ElasticsearchIndicesCheck` - check ES indices settings.
+* `\FiveLab\Component\Diagnostic\Check\Elasticsearch\ElasticsearchIndexCheck` - check ES index.